Collapse getFormValues into formAction module since it is not used anywhere else
